#41 style(be): MinutePriceDto -> MinutePriceResponse
